// frontend/src/api/generated/schema.d.ts
export interface components {
    schemas: {
// ...
        ItemResponse: {
            /** Brand */
            brand: string | null;
            /** Category Id */
            category_id: number;
            /** Climate Ids */
            climate_ids?: number[];
            /** Color */
            color: string | null;
            /** Condition Id */
            condition_id: number | null;
            /**
             * Created At
             * Format: date-time
             */
            created_at: string;
            /** Extra Attributes */
            extra_attributes: {
                [key: string]: components["schemas"]["JsonValue"];
            };
            /** Id */
            id: number;
            /** Is Active */
            is_active: boolean;
            /** Material */
            material: string | null;
            /** Model */
            model: string | null;
            /** Name */
            name: string;
            /** Notes */
            notes: string | null;
            /** Purpose Ids */
            purpose_ids?: number[];
            /** Quantity */
            quantity: number;
            /** Size */
            size: string | null;
            /**
             * Tracking Mode
             * @enum {string}
             */
            tracking_mode: "individual" | "grouped";
            /**
             * Updated At
             * Format: date-time
             */
            updated_at: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}